+++ python/trigonometry.cpp
@@ -13,11 +13,12 @@ namespace py = pybind11;
 
 template <class T> void bind_atan2(py::module &m) {
   m.def(
-      "atan2", [](const T &y, const T &x) { return atan2(y, x); }, py::arg("y"),
-      py::arg("x"), py::call_guard<py::gil_scoped_release>());
+      "atan2", [](const T &y, const T &x) { return atan2(y, x); },
+      py::kw_only(), py::arg("y"), py::arg("x"),
+      py::call_guard<py::gil_scoped_release>());
   m.def(
       "atan2", [](const T &y, const T &x, T &out) { return atan2(y, x, out); },
-      py::arg("y"), py::arg("x"), py::arg("out"),
+      py::kw_only(), py::arg("y"), py::arg("x"), py::arg("out"),
       py::call_guard<py::gil_scoped_release>());
 }